+++ sdk/computelimit/azure-mgmt-computelimit/azure/mgmt/computelimit/_client.py
@@ -19,7 +19,13 @@ from azure.mgmt.core.tools import get_arm_endpoints
 
 from ._configuration import ComputeLimitMgmtClientConfiguration
 from ._utils.serialization import Deserializer, Serializer
-from .operations import FeaturesOperations, GuestSubscriptionsOperations, Operations, SharedLimitsOperations
+from .operations import (
+    FeaturesOperations,
+    GuestSubscriptionsOperations,
+    Operations,
+    SharedLimitsOperations,
+    VmFamiliesOperations,
+)
 
 if TYPE_CHECKING:
     from azure.core import AzureClouds
@@ -37,6 +43,8 @@ class ComputeLimitMgmtClient:
     :vartype shared_limits: azure.mgmt.computelimit.operations.SharedLimitsOperations
     :ivar features: FeaturesOperations operations
     :vartype features: azure.mgmt.computelimit.operations.FeaturesOperations
+    :ivar vm_families: VmFamiliesOperations operations
+    :vartype vm_families: azure.mgmt.computelimit.operations.VmFamiliesOperations
     :param credential: Credential used to authenticate requests to the service. Required.
     :type credential: ~azure.core.credentials.TokenCredential
     :param subscription_id: The ID of the target subscription. The value must be an UUID. Required.
@@ -46,8 +54,8 @@ class ComputeLimitMgmtClient:
     :keyword cloud_setting: The cloud setting for which to get the ARM endpoint. Default value is
      None.
     :paramtype cloud_setting: ~azure.core.AzureClouds
-    :keyword api_version: The API version to use for this operation. Known values are "2026-03-20".
-     Default value is "2026-03-20". Note that overriding this default value may result in
+    :keyword api_version: The API version to use for this operation. Known values are "2026-04-30".
+     Default value is "2026-04-30". Note that overriding this default value may result in
      unsupported behavior.
     :paramtype api_version: str
     :keyword int polling_interval: Default waiting time between two polls for LRO operations if no
@@ -107,6 +115,7 @@ class ComputeLimitMgmtClient:
         )
         self.shared_limits = SharedLimitsOperations(self._client, self._config, self._serialize, self._deserialize)
         self.features = FeaturesOperations(self._client, self._config, self._serialize, self._deserialize)
+        self.vm_families = VmFamiliesOperations(self._client, self._config, self._serialize, self._deserialize)
 
     def send_request(self, request: HttpRequest, *, stream: bool = False, **kwargs: Any) -> HttpResponse:
         """Runs the network request through the client's chained policies.