+++ sdk/computelimit/azure-mgmt-computelimit/azure/mgmt/computelimit/models/__init__.py
@@ -30,6 +30,8 @@ from ._models import (  # type: ignore
     SharedLimit,
     SharedLimitProperties,
     SystemData,
+    VmFamily,
+    VmFamilyProperties,
 )
 
 from ._enums import (  # type: ignore
@@ -60,6 +62,8 @@ __all__ = [
     "SharedLimit",
     "SharedLimitProperties",
     "SystemData",
+    "VmFamily",
+    "VmFamilyProperties",
     "ActionType",
     "CreatedByType",
     "FeatureState",