+++ sdk/computelimit/azure-mgmt-computelimit/generated_samples/features_enable.py
@@ -37,6 +37,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: 2026-03-20/Features_Enable.json
+# x-ms-original-file: 2026-04-30/Features_Enable.json
 if __name__ == "__main__":
     main()