+++ sdk/computelimit/azure-mgmt-computelimit/generated_samples/features_list.py
@@ -37,6 +37,6 @@ def main():
         print(item)
 
 
-# x-ms-original-file: 2026-03-20/Features_List.json
+# x-ms-original-file: 2026-04-30/Features_List.json
 if __name__ == "__main__":
     main()