+++ sdk/computelimit/azure-mgmt-computelimit/generated_samples/guest_subscriptions_create.py
@@ -38,6 +38,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: 2026-03-20/GuestSubscriptions_Create.json
+# x-ms-original-file: 2026-04-30/GuestSubscriptions_Create.json
 if __name__ == "__main__":
     main()