+++ sdk/computelimit/azure-mgmt-computelimit/generated_samples/guest_subscriptions_delete.py
@@ -36,6 +36,6 @@ def main():
     )
 
 
-# x-ms-original-file: 2026-03-20/GuestSubscriptions_Delete.json
+# x-ms-original-file: 2026-04-30/GuestSubscriptions_Delete.json
 if __name__ == "__main__":
     main()